Add lat and lng

+++ src/main/java/com/strategy/api/domain/simpatizantes/Simpatizante.java
@@ -19,6 +19,6 @@ public class Simpatizante {
     private Simpatizante registeredBy;
     private String phoneNumber;
     private String whatsApp;
-    private Long lat;
-    private Long lng;
+    private Double lat;
+    private Double lng;
 }
