+++ src/main/java/com/strategy/api/web/register/dto/RegisterRequest.java
@@ -19,6 +19,6 @@ public class RegisterRequest {
     private String whatsApp;
     private Long userId;
     private SimpatizanteEntity registeredById;
-    private Long lat;
-    private Long lng;
+    private Double lat;
+    private Double lng;
 }
